Add VideoDetailsType to eBay SDK
A new file, VideoDetailsType.php, was created under the Trading Types directory. A class called VideoDetailsType was defined in this file, with properties that belong to this class.

In addition, the ItemType class in ItemType.php was updated to include a 'VideoDetails' property, which is an instance of the VideoDetailsType class.

These changes were made to support the management of video IDs in the eBay Trading SDK